Changed ranged based loops to use 'size_t'
I turned on all the warning flags for g++ and it started barking at me
for using 'int'. I guess this is the more correct way?

+++ src/console.cpp
@@ -79,7 +79,7 @@ void Console::clear(int x, int y)
 
 void Console::flush()
 {
-    for(int i = 0; i < screen_.size(); ++i)
+    for(size_t i = 0; i < screen_.size(); ++i)
     {
         screen_[i] = Glyph(' ');
     }
